refactor(bootstrap/assembly): extract tests to sibling tests_assembly.rs (#1266)
Inline `#[cfg(test)] mod tests` block in assembly.rs is moved into a sibling
file using the existing bootstrap convention (alongside `tests_config.rs`).

- src/bootstrap/assembly.rs: 407 → 300 LOC (under 400-line rule)
- src/bootstrap/tests_assembly.rs: new (~107 LOC)
- src/bootstrap/mod.rs: register `mod tests_assembly`
- register_builtin_base_type: `fn` → `pub(super) fn` so the sibling test
  module can reach it (was previously accessible via `super::*` from inline tests)

Pure refactor: zero logic changes, all 52 bootstrap tests pass.

Closes 1/70 of #1266 (after #1285, #1286, #1287).
