+++ src/bootstrap/mod.rs
@@ -6,6 +6,8 @@ mod validation;
 #[cfg(test)]
 mod test_support;
 #[cfg(test)]
+mod tests_assembly;
+#[cfg(test)]
 mod tests_config;
 
 // Shared constants accessible to child modules.
